style(lint): apply rule `import/no-cycle`
