+++ src/app/@dataflow/extra/index.ts
@@ -1,5 +1,4 @@
 export * from './users-flow';
-export * from './navigation-flow';
 export * from './current-user-flow';
 export * from './clipboard-flow';
 export * from './operations-list-extends-flow';
