+++ src/app/@dataflow/rclone/core-stats-reset-flow.ts
@@ -1,6 +1,5 @@
 import { AjaxFlowInteralNode, CombErr, FlowOutNode } from '../core';
-import { IRcloneServer } from '../extra';
-import { PostFlow } from './post-flow';
+import { IRcloneServer, PostFlow } from './post-flow';
 
 export interface CoreStatsResetFlowParamsNode {
 	group?: string;
