+++ src/app/@dataflow/rclone/operations-fsinfo-flow.ts
@@ -1,6 +1,6 @@
 import { AjaxFlowInteralNode, CombErr, FlowOutNode } from '../core';
-import { IRcloneServer, NavigationFlowOutNode } from '../extra';
-import { PostFlow } from './post-flow';
+import { NavigationFlowOutNode } from './navigation-flow';
+import { IRcloneServer, PostFlow } from './post-flow';
 
 export interface OperationsFsinfoFlowParamsNode {
 	/** a remote name string eg "drive:" */
